test: buffer_finalizer
