+++ packages/test/buffer/binding.c
@@ -52,17 +52,6 @@ static void noopDeleter(napi_env env, void* data, void* finalize_hint) {
   deleterCallCount++;
 }
 
-static void malignDeleter(napi_env env, void* data, void* finalize_hint) {
-  NODE_API_ASSERT_RETURN_VOID(env, data != NULL && strcmp(data, theText) == 0, "invalid data");
-  napi_ref finalizer_ref = (napi_ref)finalize_hint;
-  napi_value js_finalizer;
-  napi_value recv;
-  NODE_API_CALL_RETURN_VOID(env, napi_get_reference_value(env, finalizer_ref, &js_finalizer));
-  NODE_API_CALL_RETURN_VOID(env, napi_get_global(env, &recv));
-  NODE_API_CALL_RETURN_VOID(env, napi_call_function(env, recv, js_finalizer, 0, NULL, NULL));
-  NODE_API_CALL_RETURN_VOID(env, napi_delete_reference(env, finalizer_ref));
-}
-
 static napi_value newBuffer(napi_env env, napi_callback_info info) {
   napi_value theBuffer;
   char* theCopy;
@@ -150,30 +139,6 @@ static napi_value staticBuffer(napi_env env, napi_callback_info info) {
   return theBuffer;
 }
 
-static napi_value malignFinalizerBuffer(napi_env env, napi_callback_info info) {
-  size_t argc = 1;
-  napi_value args[1];
-  NODE_API_CALL(env, napi_get_cb_info(env, info, &argc, args, NULL, NULL));
-  NODE_API_ASSERT(env, argc == 1, "Wrong number of arguments");
-  napi_value finalizer = args[0];
-  napi_valuetype finalizer_valuetype;
-  NODE_API_CALL(env, napi_typeof(env, finalizer, &finalizer_valuetype));
-  NODE_API_ASSERT(env, finalizer_valuetype == napi_function, "Wrong type of first argument");
-  napi_ref finalizer_ref;
-  NODE_API_CALL(env, napi_create_reference(env, finalizer, 1, &finalizer_ref));
-
-  napi_value theBuffer;
-  NODE_API_CALL(
-      env,
-      napi_create_external_buffer(env,
-                                  sizeof(theText),
-                                  (void*)theText,
-                                  malignDeleter,
-                                  finalizer_ref,  // finalize_hint
-                                  &theBuffer));
-  return theBuffer;
-}
-
 static napi_value getMemoryDataAsArray(napi_env env, napi_callback_info info) {
   size_t argc = 1;
   napi_value args[1];
@@ -224,7 +189,6 @@ static napi_value Init(napi_env env, napi_value exports) {
     DECLARE_NODE_API_PROPERTY("bufferHasInstance", bufferHasInstance),
     DECLARE_NODE_API_PROPERTY("bufferInfo", bufferInfo),
     DECLARE_NODE_API_PROPERTY("staticBuffer", staticBuffer),
-    DECLARE_NODE_API_PROPERTY("malignFinalizerBuffer", malignFinalizerBuffer),
     DECLARE_NODE_API_PROPERTY("getMemoryDataAsArray", getMemoryDataAsArray),
   };
